requirements.txt
requests==2.27.1
stellar-base-sseclient==0.0.21
lxml==4.8.0
wheezy.template==3.1.0
rdflib==6.1.1
pyshacl==0.19.0